requirements.txt
# ---- Core ----
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.39.0
sentencepiece  # cần cho một số tokenizer HF

# ---- OCR ----
vietocr>=0.3.10

# ---- Vector DB ----
faiss-cpu>=1.7.4

# ---- API ----
fastapi>=0.110
uvicorn[standard]>=0.27

# ---- Utils ----
numpy==1.26.4    # Không lên 2.x để tránh crash với TF/ml-dtypes
pillow>=10.0.0
tqdm